requirements.txt
discord.py @ git+https://github.com/Rapptz/discord.py@caf0fbdc1b1e23694d440abd1feeaeacf913e845
mariadb==1.0.11
PyYAML==6.0
requests==2.28.2
vkbottle==4.3.12
revChatGPT==2.3.4

